add class support
